+++ review1/src/main/java/com/travel/review1/dao/ReviewDao.java
@@ -11,34 +11,34 @@ import com.travel.review1.repository.ReviewRepository;
 
 @Component
 public class ReviewDao {
-	@Autowired
+    @Autowired
     private ReviewRepository reviewRepository;
- 
+
     public Review saveReview(Review review) {
-    	return this.reviewRepository.save(review);
+        return this.reviewRepository.save(review);
     }
- 
-    public Optional<Review> getReviewById(Integer reviewId) {
+
+    public Optional<Review> getReviewById(Long reviewId) { // Changed from Integer to Long
         return this.reviewRepository.findById(reviewId);
     }
- 
-    public  List<Review> getAllReviews() {
+
+    public List<Review> getAllReviews() {
         return this.reviewRepository.findAll();
     }
- 
-    public List<Review> getReviewsByUserId(Integer userId) {
+
+    public List<Review> getReviewsByUserId(Long userId) { // Already Long, keeping it
         return this.reviewRepository.findByUserId(userId);
     }
- 
-    public List<Review> getReviewsByHotelId(Integer hotelId) {
+
+    public List<Review> getReviewsByHotelId(Long hotelId) { // Changed from Integer to Long
         return this.reviewRepository.findByHotelId(hotelId);
     }
- 
-    public List<Review> getReviewsByFlightId(Integer flightId) {
+
+    public List<Review> getReviewsByFlightId(Long flightId) { // Changed from Integer to Long
         return this.reviewRepository.findByFlightId(flightId);
     }
- 
-    public boolean existsById(Integer reviewId) {
+
+    public boolean existsById(Long reviewId) { // Changed from Integer to Long
         return this.reviewRepository.existsById(reviewId);
     }
